Initial Updates
Copied the original FE Project I had and cleaned it up of all our unique
references. The FE Project itself will end up on Github eventually, but
not until its completely done, and the primary use for it then would be
for allowing others to edit it to create their own games.