Update License.txt
Forgot to add some things to the License. I added in female Thief sprites which weren't added into the required License. Didn't notice the license on the Halberdier sprites, either; so much for my idea of making a female Halberdier variant. :P (Well, I could create it, but I couldn't share it with you.)